// personal/dashboard/src/lib/types.ts
export interface Contact {
  id: string;
  user_id: string;
  label_id?: string;
  name: string;
  phone_number: string;
  note?: string;
  additional_data?: any;
  created_at: string;
}

export interface CreateContactRequest {
  name: string;
  phone_number: string;
  label_id?: string;
  note?: string;
  additional_data?: Record<string, any>;
}

export interface UpdateContactRequest {
  name?: string;
  phone_number?: string;
  label_id?: string;
  note?: string;
  additional_data?: Record<string, any>;
}

export interface ContactGroup {
  id: string;
  user_id: string;
  name: string;
  description?: string;
  created_at: string;
}